Added cmake

+++ src/brush/model/geometry.cpp
@@ -1,7 +1,9 @@
 #include "geometry.h"
 #include <iostream>
-glm::vec3 Geometry::getPointFromUVCoordinates(const std::vector<glm::vec2>& pointsUV, const std::vector<glm::vec3>& points,
-                                              const glm::vec2& pointUV) {
+
+glm::vec3
+Geometry::getPointFromUVCoordinates(const std::vector<glm::vec2> &pointsUV, const std::vector<glm::vec3> &points,
+                                    const glm::vec2 &pointUV) {
     glm::vec2 uvVector1 = pointsUV[1] - pointsUV[0];
     glm::vec2 uvVector2 = pointsUV[2] - pointsUV[0];
     glm::vec3 vector1 = points[1] - points[0];
@@ -15,23 +17,24 @@ glm::vec3 Geometry::getPointFromUVCoordinates(const std::vector<glm::vec2>& poin
     return points[0] + c1 * vector1 + c2 * vector2;
 }
 
-glm::i32vec2 Geometry::toScreenCoordinates(const glm::vec3& point, const glm::mat4x4& matrixProjection, const glm::i32vec2& screenSize) {
+glm::i32vec2 Geometry::toScreenCoordinates(const glm::vec3 &point, const glm::mat4x4 &matrixProjection,
+                                           const glm::i32vec2 &screenSize) {
     glm::vec4 homogeneousCoordinates(matrixProjection * glm::vec4(point, 1.0));
     glm::vec3 projectedPoint(glm::vec3(homogeneousCoordinates.x, homogeneousCoordinates.y, homogeneousCoordinates.z)
                              / homogeneousCoordinates.w);
     return glm::i32vec2(screenSize.x * (projectedPoint.x + 1) / 2,
-                  screenSize.y * (1 - projectedPoint.y) / 2);
+                        screenSize.y * (1 - projectedPoint.y) / 2);
 }
 
-glm::vec3 Geometry::fromScreenCoordinates(const glm::vec2& point, const glm::mat4x4& matrixProjection) {
+glm::vec3 Geometry::fromScreenCoordinates(const glm::vec2 &point, const glm::mat4x4 &matrixProjection) {
     return glm::vec3(point.x / matrixProjection[0][0],
                      point.y / matrixProjection[1][1],
                      -1.0);
 }
 
-float_t Geometry::calculateFaceAngleCos(const Face& face, const glm::mat4x4& matrixModelView) {
+float_t Geometry::calculateFaceAngleCos(const Face &face, const glm::mat4x4 &matrixModelView) {
     std::vector<glm::vec3> points(3);
-    for (int i = 0; i < 3; ++i) {
+    for (size_t i = 0; i < 3; ++i) {
         points[i] = glm::vec3(matrixModelView * glm::vec4(face.getPosition(i), 1.0));
     }
     glm::vec3 normal = glm::cross(points[1] - points[0], points[2] - points[0]);
