+++ src/brush/model/objectmodel.h
@@ -8,17 +8,24 @@
 class ObjectModel {
 public:
     ObjectModel();
-    ObjectModel(uint32_t verticesNumber, uint32_t facesNumber);
-    glm::vec3 getPosition(uint32_t vertexId) const;
-    glm::vec2 getUv(uint32_t vertexId) const;
-    uint32_t getVerticesNumber() const;
-    uint32_t getFacesNumber() const;
-    void setVertex(size_t id, const glm::vec3& vertexCoordinates, const glm::vec2& vertexUv);
-    bool areAdjacentFaces(uint32_t idFace1, uint32_t idFace2) const;
+
+    ObjectModel(size_t verticesNumber, size_t facesNumber);
+
+    glm::vec3 getPosition(size_t vertexId) const;
+
+    glm::vec2 getUv(size_t vertexId) const;
+
+    size_t getVerticesNumber() const;
+
+    size_t getFacesNumber() const;
+
+    void setVertex(size_t id, const glm::vec3 &vertexCoordinates, const glm::vec2 &vertexUv);
+
+    bool areAdjacentFaces(size_t idFace1, size_t idFace2) const;
 
 private:
-    uint32_t verticesNumber_;
-    uint32_t facesNumber_;
+    size_t verticesNumber_;
+    size_t facesNumber_;
     std::vector<glm::vec3> positions_;
     std::vector<glm::vec2> uv_;
 };
